Version 27. Generates collated csv file. Adds function documentation
This commit adds a dependency in the script, on pandas, because
it is my preferred way of working with datasets. (Pandas is used to
merge the the separate datasets for each individual csv file into
a single dataset for the collated csv file.)

Also:
Fixes bugs where the "Add section here" and "Delete this section"
buttons did not always do anything, or updated a Random section but
not any depending Dependent sections.
To do so, the parseTemplateCopy function now keeps track of the
types of the sections that have been parsed.  Dependent sections do
not have those buttons...they are dependent on the Random section
so it should be changed instead.

The default repeating parameters (start, end, interval) are
usable (0, 1, 1), and added a warning in the GUI and the script
if the interval is zero.

Adds two examples for factorial designs to the webpage.

Refactors the repeatNever branches in the getChoiceFor... section
of the script, for clarity.

Adds the date/time of file creation to the csv files.

Fixes the year in the copyright notice...it should be the year
of publication/development and not be updated unless there are
significant changes.