Fix bug involving MatchDifferent and repeat chance
MatchDifferent was removing options from the
available list (freeToChoose), then "Non-uniform
chance for immediate repeat" was assuming the
previously selected choice was available and trying
to remove it, but it wasn't in the list, which
caused a crash.  Now the code checks to see if the
previously selected choice is in the list first.